kernel fast mutex: add and fix all related
1. Renamed sceKernelDeleteFastMutex to sceKernelFinalizeFastMutex. This
matches the name found in 0.940 prototype firmware and the NID has not
changed since then.

2. Unlike lightweight mutex, the work area is not required to be 8 byte
aligned, but I have done so anyway.

3. The work area's size is 0x40, but only 0x28 is used. However it is
not safe to use a smaller work area because sceKernelFinalizeFastMutex
memsets 0x40 bytes to 0.

4. The last field of the work area, 4 bytes long at offset 0x24,
contains the dummy value 0xA35DB473 when the fast mutex is initialised.
Functions in SceKernelThreadmgr related to fast mutex will check this
value and this allows us to find all functions related to fast mutex.