Add and rename old SDK functions
The suffix _16XX is used because _sceKernelCreateSema_16XX is an
official name found in syslibtrace. All the _16XX functions existed in
SDK 0.945.040 and did not exist in SDK 3.300.031. It is not necessary
that they are removed immediately after 16XX, although many of them
were.

sceKernelPuts is from SDK 0.945.040 but in system software 3.60, this
function will not do anything except return to caller.